fixing project layout, again
