+++ .idea/scala_compiler.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ScalaCompilerConfiguration">
+    <profile name="sbt 1" modules="com.microsoft.azure.cosmosdb.kafka.connect,kafka-connect-cosmosdb" />
+  </component>
+</project>
