+++ src/main/scala/com/microsoft/azure/cosmosdb/kafka/connect/config/CosmosDBConfig.scala
@@ -1,4 +1,4 @@
-package com.microsoft.azure.cosmosdb.kafka.connect.sink.config
+package com.microsoft.azure.cosmosdb.kafka.connect.config
 
 import java.util
