+++ src/main/scala/com/microsoft/azure/cosmosdb/kafka/connect/source/CosmosDBSourceTask.scala
@@ -1,0 +1,5 @@
+package com.microsoft.azure.cosmosdb.kafka.connect.source
+
+class CosmosDBSourceTask {
+
+}
