--- src/main/scala/com/microsoft/azure/cosmosdb/kafka/connect/source/README.md
@@ -1,1 +1,0 @@
-scala src for source
